Problems:
- App crashes during the filtering process.
- Image centering is still a problem.

TODOS (In order of importance):
- Implement Help button and interactive tutorial.
- Add support for large screen in the form of a landscape layout.
- Add Co-Occurrence window and spatial deviation (?).
- Implement any other TODO inside the code.

Nice to Have:
  - Change the scribble points color. (Sources: https://github.com/QuadFlask/colorpicker, https://github.com/jaredrummler/ColorPicker, https://stackoverflow.com/questions/6980906/android-color-picker)
  - Video filtering.

Things to consider in the future:
- Need to add threads that handle all the event listening and things like that because the emulator notified me that the application main thread may be doing
   too much work as the emulation skipped 42 frames (It happened more than once).
- Right now, fragments do not save state, in theory if memory is low that could prove troublesome, for example if we activated scribble and added points then continued and filtered but in the end discarded the result we might find that the button fragment from before the filtering has reset (same goes for the image viewing fragment and any other local and data that may be released) because it was re-created instead of resume due to unused memory getting released by the system.

+++ app/src/main/java/cofproject/tau/android/cof/SigmaValueWatcher.java
@@ -18,7 +18,7 @@ import static cofproject.tau.android.cof.Utility.mapSigmaToProgress;
 
 public class SigmaValueWatcher implements TextWatcher
 {
-    private Double currentVal;
+    private Float currentVal;
     private SeekBar mSigmaSeekBar;
     private EditText mTarget;
 
@@ -34,7 +34,7 @@ public class SigmaValueWatcher implements TextWatcher
     public void beforeTextChanged(CharSequence s, int start, int count, int after) {
         try
         {
-            currentVal = Double.parseDouble(s.toString());
+            currentVal = Float.parseFloat(s.toString());
         }
         catch (NumberFormatException numberFormatEx)
         {
@@ -47,12 +47,12 @@ public class SigmaValueWatcher implements TextWatcher
     @Override
     public void afterTextChanged(Editable s)
     {
-        Double newValue;
+        Float newValue;
         Integer seekBarProgress;
         boolean needToSetString = false;
         try
         {
-            newValue = Double.parseDouble(s.toString());
+            newValue = Float.parseFloat(s.toString());
             if (currentVal.equals(newValue))
             {
                 return;
